Lesson 12 (version 4.8.1). Saving intermediate results.
