+++ src/shims-vuex.d.ts
@@ -1,0 +1,7 @@
+import {store} from '@/store';
+
+declare module '@vue/runtime-core' {
+  interface ComponentCustomProperties {
+    $store: store;
+  }
+}
