+++ src/store/modules/messages.js
@@ -21,7 +21,23 @@ const store = {
             },
         ],
     },
-    getters: {},
+    getters: {
+        getMessageList: (state) => {
+            return state.MessageList;
+        },
+        getImageList: (state, getters) => {
+            const imgList = [];
+            // const imgList: Array<string> = [];
+            for (const msgItem of getters.getMessageList) {
+                if ('imageList' in msgItem) {
+                    for (const imgItem of msgItem.imageList) {
+                        imgList.push(imgItem);
+                    }
+                }
+            }
+            return imgList;
+        },
+    },
     mutations: {},
     actions: {},
 };
