+++ src/views/Calendar.vue
@@ -6,6 +6,7 @@ Calendar(:attributes="attributes" title-position="left" is-expanded )
 
 <script lang="ts">
 import {defineComponent} from 'vue';
+import {TodoInterface} from '@/types/task.interface';
 import {Calendar, DatePicker} from 'v-calendar';
 import {mapState} from 'vuex';
 
@@ -21,35 +22,33 @@ export default defineComponent({
     DatePicker,
   },
 
-  emits: {
-    'change-notifis': null,
-  },
+  emits: {},
 
   computed: {
-    ...mapState(['todoList']),
+    ...mapState('todos', {todoList: 'todoList'}),
 
     attributes() {
       return [
-        ...this.todoList.map((todo) => ({
-          dates: todo.completionDate,
-          dot: false,
-          popover: {
-            label: todo.name,
-            visibility: 'click',
-            // hideIndicator: true,
-            style: {
-              backgroundColor: 'red',
-            },
-          },
-          // customData: todo,
-          highlight: {
-            color: 'indigo',
-            fillMode: 'solid',
-            style: {
-              borderRadius: 0,
-            },
-          },
-        })),
+        // ...this.$store.modules.todos.state.todoList.map((todo: TodoInterface) => ({
+        //   dates: todo.completionDate,
+        //   dot: false,
+        //   popover: {
+        //     label: todo.name,
+        //     visibility: 'click',
+        //     // hideIndicator: true,
+        //     style: {
+        //       backgroundColor: 'red',
+        //     },
+        //   },
+        //   // customData: todo,
+        //   highlight: {
+        //     color: 'indigo',
+        //     fillMode: 'solid',
+        //     style: {
+        //       borderRadius: 0,
+        //     },
+        //   },
+        // })),
       ];
     },
   },
